refactor module renderer and add slick carousel
