+++ package.json
@@ -61,6 +61,7 @@
     "react-redux": "7.2.0",
     "react-router-config": "^5.1.1",
     "react-router-dom": "^5.1.2",
+    "react-slick": "^0.25.2",
     "redux": "3.7.2",
     "redux-thunk": "2.2.0",
     "reselect": "^4.0.0",
@@ -68,6 +69,7 @@
     "sass-loader": "^8.0.2",
     "serialize-javascript": "^3.0.0",
     "serverless-http": "^2.3.2",
+    "slick-carousel": "^1.8.1",
     "source-map-support": "^0.5.17",
     "style-loader": "^1.1.4",
     "terser-webpack-plugin": "^2.3.6",
@@ -98,6 +100,7 @@
     "@types/react-router": "5.1.4",
     "@types/react-router-dom": "5.1.3",
     "@types/sumo-logger": "1.6.1",
+    "@types/react-slick": "^0.23.4",
     "@typescript-eslint/eslint-plugin": "2.24.0",
     "@typescript-eslint/parser": "2.24.0",
     "babel-jest": "^25.4.0",
